상품 업로드 ~ing

+++ src/main/java/com/modoodesigner/domain/application/commands/PendantRegisterCommand.java
@@ -18,8 +18,8 @@ public class PendantRegisterCommand extends UserCommand {
     private int buyPrice;
     private int stock;
     private String[] tags;
-    private Long[] mainImageIndex;
-    private Long[] contentImage;
+    private int[] mainImageIndex;
+    private int[] contentImage;
     private String content;
 
 }
